+++ Tortuga.Chain/Changelog.md
@@ -4,6 +4,19 @@
 
 [#555 If a SQL Server column is marked as 'Always Generated', set the IsComputed property to true](https://github.com/TortugaResearch/Tortuga.Chain/issues/555)
 
+[#553 Find table by object id in SQL Server](https://github.com/TortugaResearch/Tortuga.Chain/issues/553)
+
+
+* Add property `ObjectId` to `SqlServerTableOrViewMetadata`
+* Add method `SqlServerObjectName GetTableOrViewName(int objectId)` to `SqlServerMetadataCache`
+* Add method `SqlServerTableOrViewMetadata<SqlDbType> GetTableOrView(int objectId)` to `SqlServerMetadataCache`
+
+
+[#552 Table Metadata should indicate if a table is a temporal table](https://github.com/TortugaResearch/Tortuga.Chain/issues/552)
+
+* Add property `HistoryTableName` to `SqlServerTableOrViewMetadata`
+* Add property `IsHistoryTable` to `SqlServerTableOrViewMetadata`
+* Add method `SqlServerTableOrViewMetadata GetHistoryTable()` to `SqlServerTableOrViewMetadata`
 
 
 ## Version 5.4.6